used for correct date & time
Switches the warning LED, if present. If the Raspberry cannot obtain the current date and time from the Internet, it is in 1970 at 1:00 am. Sooner or later this would result in watering, but not at the desired or specified times. This file checks (hourly, using crontab) the current time and turns the LED ON if necessary. During a restart, the LED will also light up until the file "boot.py" takes care of the LED.