useful for calibration
Serves to calibrate the sensors. To calculate the percentage value of the soil moisture sensors correctly, it is necessary to know the maximum value of the sensor. In theory, this is 1023, but it can be smaller due to increased resistance (cable length, soldered joints, etc.). Holding a soil moisture sensor in the air, the lowest resistance is determined and the highest value is output. The value must then be defined in the file "waterpi.py". With me the value is 918.